# docs/openapi.yaml
openapi: 3.0.3
info:
  title: Healthcare AI Billing API (example)
  version: 1.0.0
servers:
  - url: /
paths:
...
  /api/v1/accounts/{accountId}/summary:
    get:
      summary: Get account credit summary (for dashboard)
      parameters:
        - in: path
          name: accountId
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Account summary
          content:
            application/json:
              schema:
                type: object
                properties:
                  account_id: { type: string }
                  available_credits: { type: integer }
                  reserved_credits: { type: integer }
                  lifetime_earned: { type: integer }
                  lifetime_spent: { type: integer }
...